Thang

+++ src/modules/database/services/progressDBService.ts
@@ -5,10 +5,11 @@ import { ForbiddenException, Inject, Injectable, NotFoundException } from '@nest
 import { QueryParams, ResponseQuery } from 'src/interface/i-base-db-service';
 import { UnitDBService } from './unitDBService';
 import { TrainingDBService } from './trainingDBService';
-import { UserDBService } from './userDbService';
+import { UserDBService } from './userDBService';
 import { MAX_ITEM_QUERYS } from 'src/const';
-import { promises } from 'dns';
 import { PositionDBService } from './positionDBService';
+import { readFileSync } from 'fs';
+import { archiveConfig } from './../../../configs/configuration.config';
 
 @Injectable()
 export class ProgressDBService extends BaseDBService<Progress> {
@@ -38,7 +39,7 @@ export class ProgressDBService extends BaseDBService<Progress> {
     const { skip, limit } = query;
     const pageIndex = skip / limit + 1;
 
-    const checkPermisison = await this.unitDBService.checkUnitIsDescenants(userUnitID, unitProgress)
+    const checkPermisison = await this.unitDBService.checkUnitPermission(userUnitID, unitProgress)
     if(!checkPermisison) throw new ForbiddenException();
 
     const ans = await this.getItems({
@@ -47,6 +48,8 @@ export class ProgressDBService extends BaseDBService<Progress> {
         filter,
       },
     });
+
+    if(!ans) throw new NotFoundException();
     const items = ans.items;
 
     const unit = await this.unitDBService.getItemById(items[0].unit)
@@ -61,14 +64,30 @@ export class ProgressDBService extends BaseDBService<Progress> {
           return { object: temp.name, time: x.time }
         }))
 
-        item = {
-          ...item,
-          ...{
-            time_train_detail: new_time_train_detail,
-            unit: unit.name,
-            allElements: [...setElement]
+        if(item.time_train_detail.length > 0){
+
+          item = {
+            ...item,
+            ...{
+              time_train_detail: new_time_train_detail,
+              unit: unit.name,
+              allElements: [...setElement]
+            }
           }
+          
+        } else {
+
+          item = {
+            ...item,
+            ...{
+              time_train_detail: [],
+              unit: unit.name,
+              allElements: [...setElement]
+            }
+          }
+
         }
+       
         return item
       }) 
     )
@@ -81,14 +100,16 @@ export class ProgressDBService extends BaseDBService<Progress> {
     const progress = await this.getItemById(id);
     if(!progress) throw new NotFoundException();
 
-    const checkPermisison = await this.unitDBService.checkUnitIsDescenants(userUnitID, progress.unit)
+    const checkPermisison = await this.unitDBService.checkUnitPermission(userUnitID, progress.unit)
     if(!checkPermisison) throw new ForbiddenException();
     const training = await this.trainingDBService.getItemById(id);
+
     if(!training) throw new NotFoundException();
     const ans = {
       ...progress,
       ...{
-        evaluation: training.evaluation
+        evaluation: training.evaluation,
+        time_train_actual: training.time_train_actual
       }
     }
     return ans;
@@ -96,10 +117,10 @@ export class ProgressDBService extends BaseDBService<Progress> {
 
   async insertProgress(userUnitID: string, entity: any): Promise<any> {
 
-    const checkPermisison = await this.unitDBService.checkUnitIsDescenants(userUnitID, entity.unit)
+    const checkPermisison = await this.unitDBService.checkUnitPermission(userUnitID, entity.unit)
     if(!checkPermisison) throw new ForbiddenException();
-
     const progress =  await this.insertItem(entity);
+  
     if(progress){
 
       await this.trainingDBService.insertItem({
@@ -109,7 +130,7 @@ export class ProgressDBService extends BaseDBService<Progress> {
         week: progress.week,
         month: progress.month,
         year: progress.year,
-        unit: progress.unit
+        unit: progress.unit,
       });
     }
     return progress;
@@ -120,7 +141,7 @@ export class ProgressDBService extends BaseDBService<Progress> {
     const progress = await this.getItemById(progressID);
     if(!progress) throw new NotFoundException();
    
-    const checkPermisison = await this.unitDBService.checkUnitIsDescenants(userUnitID, progress.unit)
+    const checkPermisison = await this.unitDBService.checkUnitPermission(userUnitID, progress.unit)
     if(!checkPermisison) throw new ForbiddenException();
 
     const progressUpdated =  await this.updateItem(progressID, entity);
@@ -159,8 +180,9 @@ export class ProgressDBService extends BaseDBService<Progress> {
   async deleteProgress(userUnitID: string, progressID: string): Promise<any> {
 
     const ans = await this.getItemById(progressID);
+    if(!ans) throw new NotFoundException();
 
-    const checkPermisison = await this.unitDBService.checkUnitIsDescenants(userUnitID, ans.unit)
+    const checkPermisison = await this.unitDBService.checkUnitPermission(userUnitID, ans.unit)
     if(!checkPermisison) throw new ForbiddenException();
     await this.trainingDBService.removeItem(progressID);
     return await this.removeItem(progressID); 
@@ -210,7 +232,7 @@ export class ProgressDBService extends BaseDBService<Progress> {
     const progress = await this.getItemById(progressID);
     if(!progress) throw new NotFoundException();
 
-    const checkPermisison = await this.unitDBService.checkUnitIsDescenants(userUnitID, progress.unit);
+    const checkPermisison = await this.unitDBService.checkUnitPermission(userUnitID, progress.unit);
     if(!checkPermisison) throw new ForbiddenException();
 
     const lstObj = progress.time_train_detail.map((item:any) => {
@@ -240,5 +262,21 @@ export class ProgressDBService extends BaseDBService<Progress> {
       }))
       return elementJoinMap;
     }
+  }
+
+  async getContentFile(id: string, userUnitID: string): Promise<any> {
+
+    const progress = await this.getItemById(id);
+    if(!progress) throw new NotFoundException();
+
+    const checkPermisison = await this.unitDBService.checkUnitPermission(userUnitID, progress.unit)
+    if(!checkPermisison) throw new ForbiddenException();
+
+    const uploadFolder = archiveConfig().folder_saved;
+
+    const file = readFileSync(uploadFolder + progress.url);
+    const encodeFile = file.toString('base64');
+
+    return encodeFile;
   }
 }
